+++ src/Adapter/Cache/PsrCache.php
@@ -4,22 +4,27 @@ declare(strict_types=1);
 
 namespace Fight\Common\Adapter\Cache;
 
-use Fight\Common\Application\Cache\Exception\CacheException;
+use Fight\Common\Adapter\Cache\Psr6\Psr6Cache;
 use Fight\Common\Application\Cache\MutableCache;
 use Psr\Cache\CacheItemPoolInterface;
 use Psr\Log\LoggerInterface;
-use Throwable;
 
 /**
  * Class PsrCache
+ *
+ * @deprecated since 1.2.0, use Fight\Common\Adapter\Cache\Psr6\Psr6Cache. This compatibility path will be
+ *             removed in 2.0.0.
  */
 final readonly class PsrCache implements MutableCache
 {
+    private Psr6Cache $cache;
+
     /**
      * Constructs PsrCache
      */
-    public function __construct(private CacheItemPoolInterface $cachePool, private LoggerInterface $logger)
+    public function __construct(CacheItemPoolInterface $cachePool, LoggerInterface $logger)
     {
+        $this->cache = new Psr6Cache($cachePool, $logger);
     }
 
     /**
@@ -27,26 +32,7 @@ final readonly class PsrCache implements MutableCache
      */
     public function read(string $key, callable $loader, int $ttl): mixed
     {
-        try {
-            $cacheItem = $this->cachePool->getItem($key);
-
-            if (!$cacheItem->isHit()) {
-                $this->logger->debug(sprintf('Cache MISS: "%s"', $key));
-
-                $results = $loader();
-
-                $cacheItem->set($results);
-                $cacheItem->expiresAfter($ttl);
-
-                $this->cachePool->save($cacheItem);
-            } else {
-                $this->logger->debug(sprintf('Cache HIT: "%s"', $key));
-            }
-
-            return $cacheItem->get();
-        } catch (Throwable $throwable) {
-            throw new CacheException($throwable->getMessage(), $throwable->getCode(), $throwable);
-        }
+        return $this->cache->read($key, $loader, $ttl);
     }
 
     /**
@@ -54,11 +40,7 @@ final readonly class PsrCache implements MutableCache
      */
     public function delete(string $key): void
     {
-        try {
-            $this->cachePool->deleteItem($key);
-        } catch (Throwable $throwable) {
-            throw new CacheException($throwable->getMessage(), $throwable->getCode(), $throwable);
-        }
+        $this->cache->delete($key);
     }
 
     /**
@@ -66,10 +48,6 @@ final readonly class PsrCache implements MutableCache
      */
     public function clear(): void
     {
-        try {
-            $this->cachePool->clear();
-        } catch (Throwable $throwable) {
-            throw new CacheException($throwable->getMessage(), $throwable->getCode(), $throwable);
-        }
+        $this->cache->clear();
     }
 }